% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf.clumper.regular.r
\name{areThereLonelyOLD}
\alias{areThereLonelyOLD}
\title{Check if a Specific Number of Neighbors Exist for Habitat Cells}
\usage{
areThereLonelyOLD(matrix, NNeighbors, queensCase = TRUE, maxDistance = 5)
}
\arguments{
\item{matrix}{A binary matrix to check.}

\item{NNeighbors}{Exact number of neighbors to check for each habitat cell.}

\item{queensCase}{If TRUE, checks all 8 directions for neighbors; if FALSE, checks only 4.}

\item{maxDistance}{Maximum distance to look for neighbors.}
}
\value{
TRUE if any habitat cell has the specified number of neighbors within the given distance, FALSE otherwise.
}
\description{
This function checks if each habitat cell (specified by swfCat) in a matrix has a specific 
number of neighboring habitat cells within a given distance.
}
