% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_molderN.r
\name{check_neighbors}
\alias{check_neighbors}
\title{Check Neighbours Function}
\usage{
check_neighbors(v, graph, agriCat, NNeighbors, radius)
}
\arguments{
\item{v}{Node for check neighbors.}

\item{graph}{Graph object representing the habitat.}

\item{agriCat}{Category value in the graph to allocate.}

\item{NNeighbors}{Number of neighbors to check for each AgriCat node.}

\item{radius}{Radius within which to check for neighbors.}
}
\value{
Node if it meets the NNeighbors condition.
}
\description{
Checksfor the number of agriCat neighbors for each node within a specified radius.
}
