% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf.clumper.regular.r
\name{countNeighborsOLD}
\alias{countNeighborsOLD}
\title{Count Neighbors of a Cell Within a Radius}
\usage{
countNeighborsOLD(cell, cells, radius)
}
\arguments{
\item{cell}{The cell for which to count neighbors.}

\item{cells}{The matrix of cells to check against.}

\item{radius}{The radius within which to count neighbors.}
}
\value{
The count of neighbors within the specified radius.
}
\description{
Counts the number of neighbors a given cell has within a specified radius, 
useful for assessing habitat connectivity.
}
