% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf.clumper.regular.r
\name{distance}
\alias{distance}
\title{Calculate Euclidean Distance Between Two Points}
\usage{
distance(x1, y1, x2, y2)
}
\arguments{
\item{x1}{X-coordinate of the first point.}

\item{y1}{Y-coordinate of the first point.}

\item{x2}{X-coordinate of the second point.}

\item{y2}{Y-coordinate of the second point.}
}
\value{
Euclidean distance between the two points.
}
\description{
This function calculates the Euclidean distance between two points, typically in a matrix.
}
