% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf.clumper.regular.r
\name{findCell}
\alias{findCell}
\title{Find a Cell Based on Local Density}
\usage{
findCell(cells, radius = 5, Density = "M", deBug = FALSE)
}
\arguments{
\item{cells}{A matrix of cells to search from.}

\item{radius}{Radius within which to calculate local density.}

\item{Density}{Density selection mode: "H" for high, "L" for low, "M" for median density.}
}
\value{
Coordinates of the chosen cell based on the specified density.
}
\description{
This function finds a cell in a matrix based on the local density of a specified category 
(habitat or non-habitat). It can be set to find cells in high, low, or median density areas.
}
