% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_molderN.r
\name{GfM}
\alias{GfM}
\title{Graph from Matrix Function}
\usage{
GfM(mat, np = 1)
}
\arguments{
\item{mat}{Matrix representing the initial habitat state.}

\item{np}{Number of cores for parallel processing.}
}
\value{
Graph object corresponding to the input matrix.
}
\description{
Attempts to create a graph object from a matrix, with nodes representing cells and edges representing adjacency. I've taken this partially from StackExchange but it seems very unefficient
}
