% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_clc_swf.R
\name{load_clc_swf}
\alias{load_clc_swf}
\title{Load SWF+CLC data}
\usage{
load_clc_swf()
}
\value{
A `SpatRaster` object representing the SWF+CLC data.
}
\description{
This function loads and returns the Copernicus SWF+CLC stored within the package.
}
