% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_molderN.r
\name{MfG}
\alias{MfG}
\title{Matrix from Graph Function}
\usage{
MfG(g, nrows, ncols)
}
\arguments{
\item{g}{Graph object to convert into a matrix.}

\item{nrows}{Number of rows for the resulting matrix.}

\item{ncols}{Number of columns for the resulting matrix.}
}
\description{
Creates a matrix from a graph object, with cells representing nodes.
}
