% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf.clumper.regular.r
\name{resample}
\alias{resample}
\title{Resample Elements from a Vector}
\usage{
resample(x, ...)
}
\arguments{
\item{x}{Vector to resample from.}
}
\value{
A resampled element from the vector.
}
\description{
Resamples elements from a given vector, typically used for selecting random elements.
}
