% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clc_swf.R
\docType{data}
\name{swf_clc}
\alias{swf_clc}
\title{SWF+CLC2018}
\format{
A \code{SpatRaster} containing the following elements:
\describe{
  \item{CLC}{SWF+CLC}
}
}
\source{
\url{https://land.copernicus.eu/global/products/clc}
}
\usage{
load_swf_clc()
}
\description{
A \code{SpatRaster} (EPSG: 4326).
}
\details{
This dataset provides a ?
}
\examples{
clc_swf <- terra::unwrap(readRDS(system.file("extdata", "clc_swf.rds", package = "SWFmolder")))
}
\references{
\url{https://land.copernicus.eu/global/products/clc}
}
\keyword{datasets}
