% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swf_molderN.r
\name{swf_molderN}
\alias{swf_molderN}
\title{SWF Molder Network Function}
\usage{
swf_molderN(
  g,
  swfCat,
  agriCat,
  iterations,
  Q,
  NNeighbors = 1,
  swfCover = 0.3,
  max_radius = 1,
  np = 1
)
}
\arguments{
\item{g}{Graph object representing the habitat.}

\item{swfCat}{Category value in the graph to be increased.}

\item{agriCat}{Category value in the graph to allocate.}

\item{iterations}{Number of iterations to run the modification process.}

\item{Q}{Number of nodes to be potentially modified in each iteration.}

\item{NNeighbors}{Number of agriCat neighbor nodes to consider for allocation.}

\item{swfCover}{Desired cover proportion for the swfCat category.}

\item{max_radius}{Maximum search radius for neighbours.}

\item{np}{Number of cores for parallel processing.}
}
\description{
Attemps tomodifies a habitat graph to increase the cover of a specified category
}
